Support snapshots for the review pages
