+++ __test__/createReviewPages.test.js
@@ -1,8 +1,9 @@
 const path = require('path');
 const { rimrafSync } = require('rimraf');
 const { getFiles, getFileContent } = require('./utils');
-
 const createAllTests = require('../scripts/create-all-tests/createAllTests');
+const createReviewPages =
+  require('../scripts/test-reviewer/createReviewPages.mjs').createReviewPages;
 
 const __testDirectory__ = path.dirname(__filename);
 const mockTestsDirectory = path.join(__testDirectory__, '__mocks__', 'tests');
@@ -18,57 +19,61 @@ const defaultConfig = {
   args: defaultArgs,
 };
 
-const getBuildFiles = (testPlan, extensions = ['.html', '.json', '.js']) => {
-  const testPlanBuildDirectory = path.join(buildDirectory, 'tests', testPlan);
-  const fileNames = getFiles(testPlanBuildDirectory, extensions);
+const getReviewFiles = (testPlan, extensions = ['.html', '.json', '.js']) => {
+  const reviewBuildDirectory = path.join(buildDirectory, 'review');
+  const fileNames = getFiles(reviewBuildDirectory, extensions);
 
-  return fileNames.map(filePath => getFileContent(path.join(testPlanBuildDirectory, filePath)));
+  return [
+    // Also get build/index.html
+    getFileContent(path.join(buildDirectory, 'index.html')),
+    ...fileNames.map(filePath => getFileContent(path.join(reviewBuildDirectory, filePath))),
+  ];
 };
 
 beforeEach(() => {
   rimrafSync(buildDirectory);
 });
 
-describe('V1 test format version', () => {
-  it('runs createAllTests successfully (banner)', async () => {
+describe('review pages creation', () => {
+  it('runs createReviewPages successfully for v1 tests (banner)', async () => {
     const testplan = 'banner';
     const config = { ...defaultConfig, args: { ...defaultArgs, testplan } };
     await createAllTests(config);
+    createReviewPages(config);
 
-    const files = getBuildFiles(testplan);
+    const files = getReviewFiles(testplan);
     files.forEach(file => expect(file).toMatchSnapshot());
   });
-});
 
-describe('V2 test format version', () => {
-  it('runs createAllTests successfully (alert)', async () => {
+  it('runs createReviewPages successfully for v2 tests (alert)', async () => {
     const testplan = 'alert';
     const config = { ...defaultConfig, args: { ...defaultArgs, testplan } };
     await createAllTests(config);
+    createReviewPages(config);
 
-    const files = getBuildFiles(testplan);
+    const files = getReviewFiles(testplan);
     files.forEach(file => expect(file).toMatchSnapshot());
   });
 
-  it('runs createAllTests successfully (command-button)', async () => {
+  it('runs createReviewPages successfully for v2 tests (command-button)', async () => {
     const testplan = 'command-button';
     const config = { ...defaultConfig, args: { ...defaultArgs, testplan } };
     await createAllTests(config);
+    createReviewPages(config);
 
-    const files = getBuildFiles(testplan);
+    const files = getReviewFiles(testplan);
     files.forEach(file => expect(file).toMatchSnapshot());
   });
-});
 
-describe('all test format versions', () => {
-  it('runs createAllTests successfully', async () => {
-    const config = { ...defaultConfig, args: { ...defaultArgs, verbose: true } };
+  it('runs createReviewPages successfully for all test format versions', async () => {
+    const config = { ...defaultConfig, args: { ...defaultArgs } };
     await createAllTests(config);
+    createReviewPages(config);
 
     const files = [
-      ...getBuildFiles('alert'),
-      ...getBuildFiles('banner'),
-      ...getBuildFiles('command-button'),
+      ...getReviewFiles('alert'),
+      ...getReviewFiles('banner'),
+      ...getReviewFiles('command-button'),
     ];
     files.forEach(file => expect(file).toMatchSnapshot());
   });
