+++ lib/data/process-test-directory.js
@@ -81,9 +81,9 @@ const processTestDirectory = async config => {
   /** Name of the test plan. */
   const testPlanName = path.basename(directory);
 
-  // cwd; @param rootDirectory is dependent on this file not moving from the `lib/data` folder
-  const libDataDirectory = path.dirname(__filename);
-  const rootDirectory = path.join(libDataDirectory, '../..');
+  // @param rootDirectory is dependent on this file not moving from the `lib/data` folder
+  const cwd = path.dirname(__filename);
+  const rootDirectory = path.join(cwd, '../..');
 
   const testsDirectory = config?.testsDirectory ?? path.join(rootDirectory, 'tests');
   const testPlanDirectory = path.join(testsDirectory, testPlanName);
