+++ package.json
@@ -12,7 +12,7 @@
     "validate": "npm run create-all-tests -- --verbose --validate",
     "cleanup": "rimraf build",
     "create-all-tests": "cross-var node scripts/create-all-tests --testplan=$npm_config_testplan",
-    "review-tests": "cross-var node --experimental-modules scripts/test-reviewer.mjs --testplan=$npm_config_testplan",
+    "review-tests": "cross-var node --experimental-modules scripts/test-reviewer/index.mjs --testplan=$npm_config_testplan",
     "update-reference": "node scripts/update-reference.js",
     "make-v2": "node scripts/v2maker.js",
     "test": "jest",
@@ -31,6 +31,9 @@
     "simple-git": "^3.17.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.25.3",
+    "babel-jest": "^29.7.0",
     "eslint": "^8.35.0",
     "eslint-config-prettier": "^8.7.0",
     "eslint-plugin-import": "^2.27.5",
