+++ scripts/test-reviewer/review-template.mustache
@@ -19,7 +19,7 @@
 
     </style>
     <script type="module">
-      import {TestWindow} from '../tests/resources/aria-at-test-window.mjs';
+      import {TestWindow} from '../../tests/resources/aria-at-test-window.mjs';
 
       function numberTests(value) {
         let testNumber = 0;
